feat: #740 - people add flow

+++ src/main/java/com/skapp/community/peopleplanner/payload/request/employee/employment/EmployeeEmploymentBasicDetailsDto.java
@@ -16,7 +16,7 @@ public class EmployeeEmploymentBasicDetailsDto {
 
 	private EmploymentAllocation employmentAllocation;
 
-	private int[] teamIds;
+	private Long[] teamIds;
 
 	private EmployeeEmploymentBasicDetailsManagerDetailsDto primarySupervisor;
